+++ src/ic-cdk-macros/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Export composite_query to Candid. (#419)
+
 ## [0.7.1] - 2023-07-27
 
 ### Fixed